receiver/requirements.txt
scipy
numpy
paho-mqtt
Flask-RESTful
